+++ app/src/main/java/com/example/footballmatchmanager/RankingAsistencias.kt
@@ -63,7 +63,7 @@ class RankingAsistencias : Fragment() {
                     for (document in task.result!!) {
                         try {
                             val nombre = document.getString("nombre")
-                            val valoracion = document.getDouble("valoracion")
+                            val valoracion = document.getLong("valoracion")
                             val posicion = document.getString("posicion")
                             val goles = document.getLong("goles")
                             val asistencias = document.getLong("asistencias")
